Final able to complete a task
When you complete a task the MainMenuActivity will show a green
complete.

When you click on a task you are given the option to complete a task.
When you click the "Complete" button a dialog menu will appear if the
requirments are not met the user will be notified. If the requirments
are met then the user is able to complete the task and send a message.
They are also given the option to cancel.